Refactor protocol form and processing logic
updated user inputs to remove the channel length, etc. Made it easier for me to add new variables if needed. Updated index.html to be pure HTML because I felt like it's easier for me to change.

Next step is to update the protocol runner to process all of these inputs, and building the updated runners.